SONIC: Move the outer main loop into a class Game
